Cleans up cli tool
This should make it slightly easier
to add extra functions
